requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
binarytree==6.5.1
black==23.3.0
bleach==6.0.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==41.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20230430
dill==0.3.6
dnspython==2.3.0
docopt==0.6.2
docstring-to-markdown==0.12
docutils==0.20.1
easyocr==1.7.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
Fiona==1.9.4.post1
flake8==6.0.0
folium==0.14.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.4.0
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imagesize==1.4.1
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.6.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
jaraco.classes==3.2.3
jedi==0.18.2
jellyfish==0.11.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==24.0.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.3
Levenshtein==0.21.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
numpydoc==1.5.0
openai==0.27.10
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pandasai==1.1.3
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.1.1
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
plotly==5.15.0
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.4
pylint-venv==3.0.2
pyls-spyder==0.4.0
pymongo==4.3.3
Pympler==1.0.1
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyproj==3.6.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.21.1
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.3
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.1
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.3
QtPy==2.3.1
Quandl==3.7.0
rapidfuzz==3.2.0
regex==2023.8.8
reqs==0.0.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rope==1.8.0
rsa==4.9
Rtree==1.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
setuptools-scm==7.1.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.3
spyder-kernels==2.4.3
SQLAlchemy==2.0.16
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.26.0
streamlit-folium==0.13.0
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
tesseract==0.1.3
text-unidecode==1.3
textdistance==4.5.0
threadpoolctl==3.2.0
three-merge==0.1.1
tifffile==2023.8.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2+cu117
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
ujson==5.8.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
vega-datasets==0.9.0
Wand==0.6.11
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
whatthepatch==1.0.5
widgetsnbextension==4.0.7
wrapt==1.15.0
yapf==0.40.1
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
